Introduce ClassUtils
