+++ tests/test_analyzer.py
@@ -376,14 +376,14 @@ class ClassAnalyzerTests(ClassAnalyzerBaseTestCase):
         self.assertEqual(1, len(target.extensions))
         self.assertEqual(0, mock_copy_attributes.call_count)
 
-    @mock.patch.object(ClassAnalyzer, "flatten_duplicate_attributes")
+    @mock.patch.object(ClassAnalyzer, "merge_duplicate_attributes")
     @mock.patch.object(ClassAnalyzer, "flatten_attribute")
     @mock.patch.object(ClassAnalyzer, "expand_attribute_group")
     def test_flatten_attributes(
         self,
         mock_expand_attribute_group,
         mock_flatten_attribute,
-        mock_flatten_duplicate_attributes,
+        mock_merge_duplicate_attributes,
     ):
         target = ClassFactory.create(
             attrs=[
@@ -401,7 +401,7 @@ class ClassAnalyzerTests(ClassAnalyzerBaseTestCase):
         mock_flatten_attribute.assert_has_calls(
             [mock.call(target, attr) for attr in target.attrs]
         )
-        mock_flatten_duplicate_attributes.assert_called_once_with(target)
+        mock_merge_duplicate_attributes.assert_called_once_with(target)
 
     @mock.patch.object(ClassAnalyzer, "find_class")
     def test_expand_attribute_group(self, mock_find_class):
@@ -519,56 +519,6 @@ class ClassAnalyzerTests(ClassAnalyzerBaseTestCase):
             "Missing type implementation: %s", common.type.__name__
         )
 
-    def test_flatten_duplicate_attributes(self):
-        one = AttrFactory.create(local_type=TagType.ATTRIBUTE)
-        one_clone = one.clone()
-        restrictions = Restrictions(min_occurs=10, max_occurs=15)
-        two = AttrFactory.create(local_type=TagType.ELEMENT, restrictions=restrictions)
-        two_clone = two.clone()
-        two_clone.restrictions.min_occurs = 5
-        two_clone.restrictions.max_occurs = 5
-        two_clone_two = two.clone()
-        two_clone_two.restrictions.min_occurs = 4
-        two_clone_two.restrictions.max_occurs = 4
-        three = AttrFactory.create(local_type=TagType.ELEMENT)
-        four = AttrFactory.create(local_type=TagType.ENUMERATION)
-        four_clone = four.clone()
-        five = AttrFactory.create(local_type=TagType.ELEMENT)
-        five_clone = five.clone()
-        five_clone_two = five.clone()
-
-        target = ClassFactory.create(
-            attrs=[
-                one,
-                one_clone,
-                two,
-                two_clone,
-                two_clone_two,
-                three,
-                four,
-                four_clone,
-                five,
-                five_clone,
-                five_clone_two,
-            ]
-        )
-
-        winners = [one, two, three, four, five]
-
-        self.analyzer.flatten_duplicate_attributes(target)
-        self.assertEqual(winners, target.attrs)
-
-        self.assertIsNone(one.restrictions.min_occurs)
-        self.assertIsNone(one.restrictions.max_occurs)
-        self.assertEqual(4, two.restrictions.min_occurs)
-        self.assertEqual(24, two.restrictions.max_occurs)
-        self.assertIsNone(three.restrictions.min_occurs)
-        self.assertIsNone(three.restrictions.max_occurs)
-        self.assertIsNone(four.restrictions.min_occurs)
-        self.assertIsNone(four.restrictions.max_occurs)
-        self.assertEqual(0, five.restrictions.min_occurs)
-        self.assertEqual(3, five.restrictions.max_occurs)
-
     @mock.patch.object(ClassAnalyzer, "find_class")
     def test_flatten_enumeration_unions(self, mock_find_class):
         enum_a = ClassFactory.create(
@@ -650,89 +600,3 @@ class ClassAnalyzerTests(ClassAnalyzerBaseTestCase):
         self.assertEqual(1, len(classes))
         self.assertEqual(1, len(classes[0].extensions))
         self.assertEqual(expected, classes[0].extensions[0].restrictions.asdict())
-
-    def test_create_default_attribute(self):
-        extension = ExtensionFactory.create()
-        item = ClassFactory.create(extensions=[extension])
-
-        ClassAnalyzer.create_default_attribute(item, extension)
-        expected = AttrFactory.create(
-            name="value",
-            index=0,
-            default=None,
-            types=[extension.type],
-            local_type=TagType.EXTENSION,
-        )
-
-        self.assertEqual(1, len(item.attrs))
-        self.assertEqual(0, len(item.extensions))
-        self.assertEqual(expected, item.attrs[0])
-
-    def test_create_default_attribute_with_any_type(self):
-        extension = ExtensionFactory.create(
-            type=AttrTypeFactory.create(name=DataType.ANY_TYPE.code, native=True),
-            restrictions=Restrictions(min_occurs=1, max_occurs=1, required=True),
-        )
-        item = ClassFactory.create(extensions=[extension])
-
-        ClassAnalyzer.create_default_attribute(item, extension)
-        expected = AttrFactory.create(
-            name="##any_element",
-            index=0,
-            wildcard=True,
-            default=None,
-            types=[extension.type.clone()],
-            local_type=TagType.ANY,
-            restrictions=Restrictions(min_occurs=1, max_occurs=1, required=True),
-        )
-
-        self.assertEqual(1, len(item.attrs))
-        self.assertEqual(0, len(item.extensions))
-        self.assertEqual(expected, item.attrs[0])
-
-    @mock.patch.object(ClassAnalyzer, "copy_inner_classes")
-    @mock.patch.object(ClassAnalyzer, "clone_attribute")
-    def test_copy_attributes(self, mock_clone_attribute, mock_copy_inner_classes):
-        mock_clone_attribute.side_effect = lambda x, y, z: x.clone()
-        target = ClassFactory.create(
-            attrs=[AttrFactory.create(name="foo:a"), AttrFactory.create(name="b")]
-        )
-        source = ClassFactory.create(
-            attrs=[
-                AttrFactory.create(name="c"),
-                AttrFactory.create(name="a"),
-                AttrFactory.create(name="boo:b"),
-                AttrFactory.create(name="d"),
-            ]
-        )
-        extension = ExtensionFactory.create(type=AttrTypeFactory.create(name="foo:foo"))
-        target.extensions.append(extension)
-
-        self.analyzer.copy_attributes(source, target, extension)
-
-        self.assertEqual(["c", "foo:a", "b", "d"], [attr.name for attr in target.attrs])
-        mock_copy_inner_classes.assert_called_once_with(source, target)
-        mock_clone_attribute.assert_has_calls(
-            [
-                mock.call(source.attrs[0], extension.restrictions, "foo"),
-                mock.call(source.attrs[3], extension.restrictions, "foo"),
-            ]
-        )
-
-    def test_clone_attribute(self):
-        attr = AttrFactory.create(
-            restrictions=RestrictionsFactory.create(length=1),
-            types=[
-                AttrTypeFactory.create(name="foo:x"),
-                AttrTypeFactory.create(name="y"),
-                AttrTypeFactory.create(name="z", native=True),
-            ],
-        )
-        restrictions = RestrictionsFactory.create(length=2)
-        prefix = "foo"
-
-        clone = self.analyzer.clone_attribute(attr, restrictions, prefix)
-
-        self.assertEqual(["foo:x", "foo:y", "z"], [x.name for x in clone.types])
-        self.assertEqual(2, clone.restrictions.length)
-        self.assertIsNot(attr, clone)
