+++ xsdata/analyzer.py
@@ -13,10 +13,9 @@ from xsdata.models.codegen import Attr
 from xsdata.models.codegen import AttrType
 from xsdata.models.codegen import Class
 from xsdata.models.codegen import Extension
-from xsdata.models.codegen import Restrictions
 from xsdata.models.enums import DataType
-from xsdata.models.enums import TagType
 from xsdata.utils import text
+from xsdata.utils.codegen import ClassUtils
 
 
 def simple_type(item: Class):
@@ -30,7 +29,7 @@ class AttributeComparison(Enum):
 
 
 @dataclass
-class ClassAnalyzer:
+class ClassAnalyzer(ClassUtils):
     """
     Class analyzer is responsible to minize the final classes footprint by
     merging and flattening extensions and attributes.
@@ -284,7 +283,7 @@ class ClassAnalyzer:
         for attr in list(target.attrs):
             self.flatten_attribute(target, attr)
 
-        self.flatten_duplicate_attributes(target)
+        self.merge_duplicate_attributes(target)
 
     def expand_attribute_group(self, target: Class, attr: Attr):
         """
@@ -346,90 +345,3 @@ class ClassAnalyzer:
     def is_attribute_self_reference(self, target: Class, dependency: AttrType) -> bool:
         qname = target.source_qname(dependency.name)
         return self.find_class(qname, condition=lambda x: x is target) is not None
-
-    @classmethod
-    def flatten_duplicate_attributes(cls, target: Class):
-        """
-        Flatten duplicate attributes.
-
-        Remove duplicate fields in case of attributes or enumerations
-        otherwise convert fields to lists.
-        """
-        result: List[Attr] = []
-        for attr in target.attrs:
-            existing = next((item for item in result if attr == item), None)
-
-            if not existing:
-                result.append(attr)
-            elif not (attr.is_attribute or attr.is_enumeration):
-                min_occurs = existing.restrictions.min_occurs or 0
-                max_occurs = existing.restrictions.max_occurs or 1
-                attr_min_occurs = attr.restrictions.min_occurs or 0
-                attr_max_occurs = attr.restrictions.max_occurs or 1
-
-                existing.restrictions.min_occurs = min(min_occurs, attr_min_occurs)
-                existing.restrictions.max_occurs = max_occurs + attr_max_occurs
-
-        target.attrs = result
-
-    @classmethod
-    def copy_attributes(cls, source: Class, target: Class, extension: Extension):
-        prefix = text.prefix(extension.type.name)
-        target.extensions.remove(extension)
-        target_attr_names = {text.suffix(attr.name) for attr in target.attrs}
-
-        index = 0
-        for attr in source.attrs:
-            if text.suffix(attr.name) not in target_attr_names:
-                clone = cls.clone_attribute(attr, extension.restrictions, prefix)
-                target.attrs.insert(index, clone)
-            index += 1
-
-        cls.copy_inner_classes(source, target)
-
-    @classmethod
-    def clone_attribute(
-        cls, attr: Attr, restrictions: Restrictions, prefix: Optional[str] = None
-    ):
-        clone = attr.clone()
-        clone.restrictions.update(restrictions)
-        if prefix:
-            for attr_type in clone.types:
-                if not attr_type.native and attr_type.name.find(":") == -1:
-                    attr_type.name = f"{prefix}:{attr_type.name}"
-
-        return clone
-
-    @classmethod
-    def copy_inner_classes(cls, source: Class, target: Class):
-        for inner in source.inner:
-            exists = next(
-                (found for found in target.inner if found.name == inner.name), None
-            )
-            if not exists:
-                target.inner.append(inner)
-
-    @classmethod
-    def create_default_attribute(cls, item: Class, extension: Extension):
-        if extension.type.native_code == DataType.ANY_TYPE.code:
-            attr = Attr(
-                name="##any_element",
-                index=0,
-                wildcard=True,
-                default=list if extension.restrictions.is_list else None,
-                types=[extension.type.clone()],
-                local_type=TagType.ANY,
-                restrictions=extension.restrictions.clone(),
-            )
-        else:
-            attr = Attr(
-                name="value",
-                index=0,
-                default=None,
-                types=[extension.type.clone()],
-                local_type=TagType.EXTENSION,
-                restrictions=extension.restrictions.clone(),
-            )
-
-        item.attrs.insert(0, attr)
-        item.extensions.remove(extension)
